Move VisualStudioProject.BatchingDocumentCollection to it's own file
This is just a direct invocation of the refactoring; no other work
was done.
